Append a more explanatory error message for Compute Operation errors. (#6403) (#918)
* Prefer the Message field on LocalizedMessage to that of the top level Message.

* Also update go.sum

* Write full Operation error including ErrorInfo

Do not just write the Operation error's LocalizedMessage.Message, but
alos maintain the top level Message in the `Error() string` return value
and also serialize the ErrorInfo that might exists in the error response
as json to the returned error string.

* Don't need to write an additional new line

* Need encoding/json for ga too because of ErrorInfo

Since we're serializing ErrorInfo to json and appending to a buffer for
the ComputeOperationError message, we need encoding/json in ga too.

* Do not append ErrorInfo as JSON

Append LocalizedMessage.Message to Error.Message, but do not serailize
an ErrorInfo, if one were to exist, as JSON to the io.Writer.

* Need to import io for io.StringWriter

Signed-off-by: Modular Magician <magic-modules@google.com>

Signed-off-by: Modular Magician <magic-modules@google.com>